requirements.txt
# AWS CDK
aws-cdk-lib==2.100.0
constructs>=10.0.0

# ML Training
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0

